v0.1.0 - Main Page, Router, Search Bar
Added the Main and Details Page.
The main page features a search bar with input validation.
If the input is validated, the user will be redirected to the details page.
If it's not, a popup message will be displayed.

Implemented the React Router and Framer Motion for smooth animations when switching between pages.

Added general project styles (CSS variables, color palette, fonts, etc.).
Added the page favicon and logo.