v0.2.0 - Search List API Integration
Implemented the search list.
The search bar input sends a fetch request to the API to retrieve the data necessary to make a list of items below the search bar.
The user can choose an item to display the full data on the details page (another API fetch request - not yet implemented).

Moved the validation code to an external function for readability and maintainability.

The search bar will slightly move to the top to make space for the list.

Implemented minor design changes.