v0.3.1- Search Autocomplete
When the user inputs a keyword, it is validated, and then clicks enter/search button, the first element from the search list will be searched.
It improves the UX because the user doesn't have to input the entire keyword.
For example, by inputting "chickp" and pressing "enter," the system will search for "chickpeas" (assuming it's the first result in the search list).