v0.3.2 - Loading Animation, Popup Messages
Added an animated icon when the "search list" is loading and when fetching the search bar results (details).
The icon disappears after a slight delay to prevent rapid flashing between each typed character.

Added a popup message when the API takes too long to respond.

Changed the error message to be displayed in the tile instead of the search list.
The error response codes are now inside a separate function to improve readability and maintainability.

The result list is now hidden when searching for a product and pressing enter.

Implemented other minor adjustments and improvements for the popup message system.