v0.5.0 - Page Print, Closest Section
Implemented the "PDF Print" feature.
Users now have the ability to print the entire page or save it as a PDF document.
Before the window.print() function takes effect, the page is covered with a white background and resized to the print size (a trick to ensure the print is a one, long page instead of multiple A4 pages).
After the page is printed/saved as PDF, the cover fades out and the page returns to its previous appearance.

Added the "closest section" navigation feature.
The page detects currently displayed section on the page and highlights the corresponding link button in the "nav-content" section.

Now, when the user clicks outside the nav (on mobile), it will collapse.

Fixed a bug where if the user searched for an item and then clicked the search button or enter (when focused on the search bar), it would load the previous item.
It was caused by the first list element's value not being updated when clicking the search button or enter.