v0.6.0 - Favorites List
Added the Favorites List.
The user can bookmark the current search.

All bookmarked searches will be added to the favorites list (saved to local storage) and can be displayed in the nav contents section when clicking the fav list button.
The item can be removed from the list (X button) or the user can click the name of the item to search it.

The buttons will be marked to indicate which content is currently displayed in the "nav contents section".
Clicking another button will unmark the previous one, and the "nav contents section" will change its content to the one that corresponds to the clicked button.
Clicking the same button again will unmark it, and the "nav contents section" will display the default "scrollTo" page links with the closest page indicator.