v0.7.0 - Body Metrics, UX
Added the Body Metrics navigation section.
There's a brief explanation of the Mifflin St Jeor equation (an estimate of basal metabolic rate).
The user can input their data (sex, height, weight, age, and activity level) to calculate the BMR/AMR and BMI values, which will be displayed in a table.
BMI values are compared to BMI ranges, and the respective category is highlighted.
The data is saved to localStorage and a Redux store.

There is a select dropdown to choose the units - metric / imperial (US).
The calculations are performed on the superior units (metric), so the imperial units are automatically converted to them.

The inputs have to be validated to calculate the results.
These values will be used to calculate the recommended daily nutrition intake (calories, fats, micronutrients, etc) for the searched item.

Changed the "Favorites List" to display "-----" when there are no bookmarked items.
Fixed a small bug where the nav would be collapsed if the fav list item was removed (clicking the "x" button).

Changed the nav buttons' hover effect to differ from the active class design to be less confusing when unclicking the active state.
Implemented other minor design improvements.

Improved accessibility.