v0.8.0 - Multilingual Update
Added support for multiple languages:
- English
- German
- Spanish
- French
- Polish

The code automatically detects the browser's language settings and applies them to the web app.
The user can also manually switch the language.

The query (search) results and the interface are displayed in the selected language.

The favorites list is specific to the language.
For example, if the user switches from English to German, the favorite items will change to the ones previously saved when the language was set to German.
The list will be empty if no items were marked as favorite.

When switching the language, the search bar value is set to empty, and the animated placeholder takes its place.
This is done to avoid any mismatch between different item names in different languages.

Fixed an issue with lack of instant search validation.
Now, the fetch request won't be sent if the query value isn't validated.

Disabled the default Search Bar autocomplete feature.

Improved navigation on the smallest mobile devices for better appearance and to always fit the screen size.

Implemented other minor design and script changes.