v0.9.1 - 404 Page Routing Fix
Added a 404.html file to redirect to the correct 404 page (not the default GitHub Pages' 404 page).
Unfortunately, due to how GitHub Pages handles SPA routing, there is no perfect solution.
If the website isn't accessed from a link to the main page (".../keepitfit/"), but from a deeper link, for example".../keepitfit/details/", it will show the 404 page.

Fixed an issue with the homepage button link not covering the whole button on the 404 page.