v0.10.0 - Nutrition Tables, Bug Fixes
===== TABLES =====
Added the nutrition tables.
There are two tables - macronutrients and micronutrients.
The user can see the values per 100g or portion.
The portion value can be changed using the "serving size" input.

The "% Daily Value" columns are based on the user's AMR value calculated in the Body Metrics section (navigation).
If not calculated, the values are based on the default 2000 kcal daily needs.
Some of the micronutrients are static - not based on the AMR value.

===== OTHER =====
When the user clicks the "calculate" button (Navigation - Body Metrics), they will be scrolled down to the results (AMR, BMI).

Fixed a bug in the select dropdown for user's sex (Navigation - Body Metrics).
When clicking the "calculate" button with no selected option, the border would have become red (as expected).
Then, when the user did choose an option and clicked "calculate" again, the error border would not disappear.

Fixed a bug with the userAMR NaN value on the first website load.