v0.10.1 - Design & Table DV Info
Implemented the placeholder animation to the table cells when there is no content.

Added a popup message when hovering over the "% Daily Value" table cell.
The message contains information about how the values are based on the user's AMR and that they are only estimates.

Changed the "::selection" color on certain elements to use the accent color to better match the design.

Fixed a bug where the translation language would be incorrectly set after clearing localStorage.

Implemented other minor design improvements.