v0.11.0 - Nutrition Pie Charts & Bug Fixes
Added Nutrition Charts for weight and calories.
They are a visual representation of the nutrient value as a part of the whole food item.
The charts have a loading animation when they appear in page view.

Added the DV info popup in the other column and adjusted its position on mobile to fit the screen size better.

Added correct links to sections in the navigation.
Slightly adjusted the top scrolling padding height to improve "current section" indication.

Fixed a minor bug in the mobile navigation that caused an invisible space below, preventing other elements from being clicked.

Made minor changes to the nutrition table's styles so the paddings are even and the content always fits in 1 line.

Made other minor design changes.