v0.11.1 - PDF Styles
Added PDF styles for printing the page or downloading PDF.

Changed the way PDF print works to ensure the correct print height.
The updated process for printing the page without media queries:
1. Copying the original stylesheet (1x for "no media" styles, 1x for the original styles)
2. Removing the media queries from 1 copy
3. Removing all of the styles and replacing them with the "no media" copy
4. Printing the page
5. Removing the "no media" styles and reapplying the original styles

This ensures that mobile styling has no effect on the print view.

Fixed an issue with the net carbs in the pie charts not rounding decimals correctly.

Prevented users from printing the page if the search value is empty.
Added an indicator to the Bookmark and PDF buttons in the navigation when the action cannot be performed.
The icon stroke will briefly flash the error color.