v1.0.0 - Diet Labels, Readme
Added the Diet Labels section (translated into all available languages).
Not every label fetched from the API is used because there are a lot of repetitions and useless ones.
If there are no labels fetched from the API, the section is not displayed along with the corresponding navigation section link.

Changed the navigation "current section" detection code to indicate the Diet Labels section correctly.
The section is short and it's close to the bottom of the page, so detecting the section position in relation to the top edge of the screen would not work.

Added content to the Readme file.

Implemented minor design changes.