v1.0.3 - Pie Charts Fix
Fixed a bug in the chart's net carbs & dietary fiber values.
Previously, if there was no dietary fiber, the net carbs value would be set to 0, even if there was a value fetched from the API server.

Now, the chart gap is correctly showing up when printing the page (if the charts section was not intersected before printing).